# docs/swagger/openapi-converted.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "API for managing community watch shifts, bookings, and reports",
        "title": "Community Watch Shift Scheduler API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/auth/dev-login": {
            "post": {
                "description": "Bypasses OTP and directly generates JWT token for testing (DEV MODE ONLY)",
                "tags": [
                    "auth"
                ],
                "summary": "Development-only direct login endpoint",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/api.DevLoginRequest"
                            }
                        }
                    },
                    "description": "Phone number for dev login",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "JWT token generated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.DevLoginResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request or dev mode disabled",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Registers a new user with phone number or starts login flow for existing user",
                "tags": [
                    "auth"
                ],
                "summary": "Register a new user or request OTP for existing user",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/api.RegisterRequest"
                            }
                        }
                    },
                    "description": "User registration details",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OTP sent successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.RegisterResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid phone number or request format",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/verify": {
            "post": {
                "description": "Verifies the one-time password (OTP) and returns a JWT token on success",
                "tags": [
                    "auth"
                ],
                "summary": "Verify OTP and get authentication token",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/api.VerifyRequest"
                            }
                        }
                    },
                    "description": "Verification details",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Verified successfully, returns JWT token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.VerifyResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request format",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid OTP or verification failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "http://localhost:5888"
        }
    ],
    "components": {
        "schemas": {
...
            "api.DevLoginRequest": {
                "type": "object",
                "properties": {
                    "phone": {
                        "type": "string"
                    }
                }
            },
            "api.DevLoginResponse": {
                "type": "object",
                "properties": {
                    "token": {
                        "type": "string"
                    },
                    "user": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "integer"
                            },
                            "name": {
                                "type": "string"
                            },
                            "phone": {
                                "type": "string"
                            },
                            "role": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
...
            "api.ErrorResponse": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string"
                    }
                }
            },
...
            "api.RegisterRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    }
                }
            },
            "api.RegisterResponse": {
                "type": "object",
                "properties": {
                    "dev_otp": {
                        "description": "Development only - include OTP for easier testing",
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
...
            "api.VerifyRequest": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    }
                }
            },
            "api.VerifyResponse": {
                "type": "object",
                "properties": {
                    "token": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}